+++ pkg/osvscanner/osvscanner.go
@@ -14,7 +14,6 @@ import (
 	"github.com/google/osv-scanner/internal/imodels/results"
 	"github.com/google/osv-scanner/internal/osvdev"
 	"github.com/google/osv-scanner/internal/output"
-	"github.com/google/osv-scanner/internal/semantic"
 	"github.com/google/osv-scanner/pkg/models"
 	"github.com/google/osv-scanner/pkg/osv"
 	"github.com/google/osv-scanner/pkg/reporter"
@@ -180,32 +179,6 @@ func DoScan(actions ScannerActions, r reporter.Reporter) (models.VulnerabilityRe
 	return results, nil
 }
 
-// patchPackageForRequest modifies packages before they are sent to osv.dev to
-// account for edge cases.
-func patchPackageForRequest(pkg imodels.PackageInfo) imodels.PackageInfo {
-	// Assume Go stdlib patch version as the latest version
-	//
-	// This is done because go1.20 and earlier do not support patch
-	// version in go.mod file, and will fail to build.
-	//
-	// However, if we assume patch version as .0, this will cause a lot of
-	// false positives. This compromise still allows osv-scanner to pick up
-	// when the user is using a minor version that is out-of-support.
-	if pkg.Name == "stdlib" && pkg.Ecosystem.Ecosystem == osvschema.EcosystemGo {
-		v := semantic.ParseSemverLikeVersion(pkg.Version, 3)
-		if len(v.Components) == 2 {
-			pkg.Version = fmt.Sprintf(
-				"%d.%d.%d",
-				v.Components.Fetch(0),
-				v.Components.Fetch(1),
-				9999,
-			)
-		}
-	}
-
-	return pkg
-}
-
 // TODO(V2): This will be replaced by the new client interface
 func makeRequest(
 	r reporter.Reporter,
@@ -217,7 +190,6 @@ func makeRequest(
 	var query osv.BatchedQuery
 	for _, psr := range packages {
 		p := psr.PackageInfo
-		p = patchPackageForRequest(p)
 		switch {
 		// Prefer making package requests where possible.
 		case !p.Ecosystem.IsEmpty() && p.Name != "" && p.Version != "":
